fix pictures, add conftest.py
